requirements.txt
sanic
jinja2
pyjwt
msgpack
aiohttp